+++ frontend/src/pages/profile/ui/StatsCounters/StatsCounters.tsx
@@ -1,23 +1,6 @@
 import { Row, Text, styled } from '@nextui-org/react';
 import React, { FC } from 'react';
-
-const StatCounterContainer = styled('div', {});
-
-interface StatCounterProps {
-  amount: number;
-  title: string;
-}
-
-const StatCounter: FC<StatCounterProps> = ({ amount, title }) => {
-  return (
-    <StatCounterContainer>
-      <Text css={{ fontWeight: '700', textAlign: 'center' }} size={'$xl'}>
-        {amount}
-      </Text>
-      <Text css={{ textAlign: 'center' }}>{title}</Text>
-    </StatCounterContainer>
-  );
-};
+import { StatCounter } from './StatCounter';
 
 const Wrapper = styled('div', {
   display: 'flex',
